fix(ref): added unit tests for hashnode
