+++ code/go/0chain.net/blobbercore/handler/handler.go
@@ -81,7 +81,7 @@ func SetupHandlers(r *mux.Router) {
 	r.HandleFunc("/v1/writemarker/lock/{allocation}", WithHandler(LockWriteMarker)).Methods(http.MethodPost, http.MethodOptions)
 	r.HandleFunc("/v1/writemarker/lock/{allocation}", WithHandler(UnlockWriteMarker)).Methods(http.MethodDelete, http.MethodOptions)
 
-	r.HandleFunc("/v1/refs/root/{allocation}", WithHandler(LoadRootNode)).Methods(http.MethodGet, http.MethodOptions)
+	r.HandleFunc("/v1/hashnode/root/{allocation}", WithHandler(LoadRootHashnode)).Methods(http.MethodGet, http.MethodOptions)
 }
 
 func WithReadOnlyConnection(handler common.JSONResponderF) common.JSONResponderF {
