+++ code/go/0chain.net/blobbercore/handler/handler_hashnode.go
@@ -5,10 +5,10 @@ package handler
 
 import "github.com/0chain/blobber/code/go/0chain.net/blobbercore/reference"
 
-// LoadRootNode load root node with its descendant nodes
-func LoadRootNode(ctx *Context) (interface{}, error) {
+// LoadRootHashnode load root node with its descendant nodes
+func LoadRootHashnode(ctx *Context) (interface{}, error) {
 
-	root, err := reference.LoadRootNode(ctx, ctx.AllocationTx)
+	root, err := reference.LoadRootHashnode(ctx, ctx.AllocationTx)
 	if err != nil {
 		return nil, err
 	}
